move stack submodule into executor submodule
