+++ crates/wasmi/src/engine/regmach/executor/instrs/return_.rs
@@ -3,7 +3,7 @@ use crate::{
     core::UntypedValue,
     engine::{
         bytecode::{AnyConst32, Const32, Instruction, Register, RegisterSpan, RegisterSpanIter},
-        regmach::stack::ValueStackPtr,
+        regmach::executor::stack::ValueStackPtr,
     },
 };
 use core::slice;
