+++ crates/wasmi/src/engine/regmach/executor/stack/calls.rs
@@ -10,7 +10,7 @@ use wasmi_core::TrapCode;
 use crate::{
     engine::bytecode::Instruction,
     engine::bytecode::Register,
-    engine::regmach::stack::ValueStack,
+    engine::regmach::executor::stack::ValueStack,
     engine::CompiledFunc,
     Global,
     Memory,
